Minimal main.go
