+++ internal/exitcode/exitcode.go
@@ -1,0 +1,6 @@
+package exitcode
+
+const (
+	Success = iota
+	Interrupt
+)
